requirements.txt
Flask
Pillow
geocoder
google-cloud-logging
gunicorn
ua-parser
